Sample test-code candidates based on current open file
Previously, all tests from a project where incorporated as test-code
candidates for the LLM, however, this can be slow for larger projects.
This change uses the currently open file, determines the test methods
that call code in the currently open file, and only use those as
candidates for the LLM.  As a fallback, if none could be found, all
tests from the project will be candidates.