Provide tests for a selection of method
The core algorithm is still not tested because I do not see how I could
get the required objects in the necessary state.  The unfortunate thing
about the implemented selection algorithms is that they rely on an open
editor for a normal execution, which seems to be hard to set up for a
unit test.  Any feedback and help on this is highly appreciated!